simulation of a pulse with absorbing boundary conditions

+++ fd1d/clang/test_1_2.c
@@ -11,14 +11,14 @@
 
 
 float gaussian(int t, int t0, float sigma) {
-    return exp(-0.5*(t - t0)/sigma*(t - t0)/sigma);
+    return expf(-0.5f*(t - t0)/sigma*(t - t0)/sigma);
 }
 
 
 void exfield(int t, int nx, float *ex, float *hy) {
     /* calculate the Ex field */
     for (int i = 1; i < nx; i++)
-        ex[i] += 0.5 * (hy[i-1] - hy[i]);
+        ex[i] += 0.5f * (hy[i-1] - hy[i]);
     /* put a Gaussian pulse in the middle */
     ex[nx/2] = gaussian(t, 40, 12.0f);
 }
@@ -30,7 +30,7 @@ void hyfield(int nx, float *ex, float *hy, float *bc) {
     ex[nx-1] = bc[3], bc[3] = bc[2], bc[2] = ex[nx-2];
     /* calculate the Hy field */
     for (int i = 0; i < nx-1; i++)
-        hy[i] += 0.5 * (ex[i] - ex[i+1]);
+        hy[i] += 0.5f * (ex[i] - ex[i+1]);
 }
